remove karmacounter.js
