+++ appConstants.js
@@ -1,8 +1,8 @@
 // Prod
-const EMOJI_UPVOTE_ID = "1304553163512352788";
-const EMOJI_DOWNVOTE_ID = "1304553174509817957";
-// // Dev
-// const EMOJI_UPVOTE_ID = "1305663450793508924";
-// const EMOJI_DOWNVOTE_ID = "1305663438474973237";
+// const EMOJI_UPVOTE_ID = "1304553163512352788";
+// const EMOJI_DOWNVOTE_ID = "1304553174509817957";
+// Dev
+const EMOJI_UPVOTE_ID = "1305663450793508924";
+const EMOJI_DOWNVOTE_ID = "1305663438474973237";
 
 module.exports = { EMOJI_UPVOTE_ID, EMOJI_DOWNVOTE_ID };
